refactor(backend): split helper_functions.rs into topic-based submodules
