+++ opossum_backend/src/operations/copy.rs
@@ -1,0 +1,63 @@
+use std::collections::HashSet;
+
+use actix_web::{post, web};
+use opossum_core::types::api_types::{AnalyzerItemDto, ErrorResponse, NodeInfo};
+use uuid::Uuid;
+
+use crate::{
+    app_state::{AppState, NodeCacheItem},
+    error::BackEndErrorResponse,
+};
+
+/// Copy existing nodes
+///
+/// This function copies a single or multiple already existing nodes
+#[utoipa::path(tag = "operations",
+    request_body(content = HashSet<Uuid>,
+        description = "List of Uuids of the nodes to be copied",
+        content_type = "application/json",
+    ),
+    responses(
+        (status = OK, body= NodeInfo, description = "Node successfully created", content_type="application/json"),
+        (status = BAD_REQUEST, body = ErrorResponse, description = "UUID not found", content_type="application/json")
+    )
+)]
+#[post("/copy_nodes")]
+pub(super) async fn post_copy_nodes(
+    data: web::Data<AppState>,
+    node_id: web::Json<HashSet<Uuid>>,
+) -> Result<(), BackEndErrorResponse> {
+    let mut all_nodes_found = true;
+    let node_ids_to_copy = node_id.into_inner();
+
+    // Get optic ref of node that should be copied
+    let document = data.document.lock();
+    let mut copied_nodes_set = data.node_copy_cache.lock();
+    copied_nodes_set.clear();
+
+    for id in &node_ids_to_copy {
+        if let Ok((node_ref_to_copy, _)) = document.scenery().node_recursive(*id) {
+            copied_nodes_set.push(NodeCacheItem::Optical(node_ref_to_copy));
+        } else if let Some(analyzer) = document.analyzers().get(id).cloned() {
+            // Save the DTO in cache so we retain the ID
+            copied_nodes_set.push(NodeCacheItem::Analyzer(AnalyzerItemDto {
+                id: *id,
+                info: analyzer,
+            }));
+        } else {
+            all_nodes_found = false;
+        }
+    }
+    drop(copied_nodes_set);
+    drop(document);
+
+    if all_nodes_found {
+        Ok(())
+    } else {
+        Err(BackEndErrorResponse::new(
+            404,
+            "Opossum",
+            "Some nodes could not be copied as they were not found in the document",
+        ))
+    }
+}
